feat: weather comps moved to "components/Weather"
