+++ vue/src/components/Weather/TestBarChart.vue
